backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
boto3>=1.34
pandas>=2.0
numpy>=1.26
openpyxl>=3.1
xlrd>=2.0
python-dotenv>=1.0
python-dateutil>=2.8

# ── Testing ────────────────────────────────────────────────────────────────
pytest>=7.4
pytest-django>=4.7
moto[s3]>=5.0
